Implemented TabExpandeUI class
This class allows to cycle through and render possible tab expansions.
If there is just one expansion, it's chosen.
If there are more expansions, first the common prefix is chosen, and afterwards
the TabExpander will cycle through the possible solutions, highlighting the selected
one.
TabExpansion can be aborted with Escape.

getline.cs diff might be bigger than the actual logic changes as it's style was a little
bit asjusted to the project convetions.