New approach to set values from ExecutionVisitor
I noticed that in several operations ('+=', '++', ...) the expression
to get and set the values are evaluated twice (for getting and for setting).

This is of course wrong. Therefore I started to implement the SettableExpression
class family that allow simple getting and setting of the corresponding values
and that assure that expressions aren't evaluated more than once.

This commit is the first step of refactoring the existing functionality,
the "only evaluate once" functionality will be added later.