Fixed tests to work with PS
When it comes to unloading providers there is a weird thing: They don't
get removed by Powershell if the test is executed with Nunit. It does work
though when executed as a script in a normal shell.
As I didn't find a solution to this, unloading tests stays ignored.

Also the tests now have many lines like `(Get-PSDrive | where name -eq "someName") -eq $null`
to test whether a provider or drive is still present.
This "complicated" syntax is used because these cmdlets would WriteError()s if they don't
exist. As the common ErrorAction parameter isn't implemented in Pash, yet, we use this
syntax to check it.