First ItemCmdletProviderTests, modified TestItemProvider
I noticed that PSDriveInfo property doesn't seem to be set
if the provider is only an ItemProvider. Also it seems
that we need to invoke the cmdlets with the provider
specified class, as Powershell is not able to identify
the Drive in a path if the provider is only an ItemCmdletProvider.