Removed bug in PathNaviation, simplified SessionStateScopeTests
I accidentally introduced a bug in PathNavigation by changing the
constructor's parameter types to string. I reverted that and made
that class internal to make sure it doesn't appear in any public API.

The SessionStateScopeTests used a testProvider, so drive tests work.
As the TestModule includes an unknown number of provider and default
drives, I removed it. The tests use a DummyProvider instead, which is a
much cleaner solution.