Cleanup, first support for Include/Exclude filters
Some comments and refactoring for easier understanding the source code.

Implemented IncludeExcludeFilter as a simple class to compile wildcards
from the runtimes Include/Exclude properties, with the option to ignore
them. Although the last one avoids all actual work of the filter, it
makes using it simpler as additional checks from outside can be omitted.