Implement Select-String -InputObject.
String arrays cannot be passed as the pipeline but -InputObject
can be used to find a string inside an array of strings.

Looking at https://technet.microsoft.com/en-us/library/hh849903.aspx
the behaviour when an array of strings is passed on the pipeline
versus passed using InputObject is different.

When passing an array of strings using InputObject the array is
turned into a single string with a space between each item of the
array.

When an array of strings is passed down the pipeline each item of the
array is searched individually.