Implement Select-String -Quiet
Quiet returns $true if there is a match instead of the match
information.

The Quiet parameter has different behaviour if you are searching
files compared with searching items passed along the pipeline or
via the InputObject. With files if there is no match the
Select-String command will output $false but for pipeline or
InputObject items it will output nothing.